Refactor code into the HTML reporter package.
